+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Various new unit tests.
+- Ability to pass ``kwargs`` to ``SupplyChainNetwork()``
+
 ## [0.0.3] -- 2022-06-11
 ### Added
 - Various new unit tests.